remove SinkTaskContext and replace with anonymous function
