+++ connector/src/main/java/org/astraea/connector/backup/Exporter.java
@@ -38,7 +38,6 @@ import org.astraea.common.consumer.Record;
 import org.astraea.connector.Definition;
 import org.astraea.connector.SinkConnector;
 import org.astraea.connector.SinkTask;
-import org.astraea.connector.SinkTaskContext;
 import org.astraea.connector.TaskContext;
 import org.astraea.fs.FileSystem;
 
@@ -278,7 +277,7 @@ public class Exporter extends SinkConnector {
 
       this.fs = FileSystem.of(configuration.requireString(SCHEMA_KEY.name()), configuration);
       this.writerFuture = CompletableFuture.runAsync(createWriter());
-      this.taskContext = new SinkTaskContext(context);
+      this.taskContext = TaskContext.of(this.context);
     }
 
     @Override
